+++ src/GameRandom/App.axaml.cs
@@ -1,16 +1,12 @@
 using System;
 using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
-using Avalonia.Data.Core;
 using Avalonia.Data.Core.Plugins;
 using System.Linq;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
-using GameRandom.CoreApp;
-using GameRandom.Scr.WindowScr;
+using GameRandom.Scr.DI;
 using GameRandom.SteamSDK;
-using GameRandom.SteamSDK.DI;
-using GameRandom.ViewModels;
 using GameRandom.Views;
 
 namespace GameRandom;
@@ -54,13 +50,13 @@ public partial class App : Application
 
     private void RegisterUiService(Window window)
     {
-        var factory = Di.Container.GetInstance(typeof(DiFactory)) as DiFactory;
+        var factory = Di.Container.GetInstance<DiFactory>() as DiFactory;
         
         if (factory == null)
             throw new Exception("DiFactory not found");
-        
-        if (window is MainWindow mainWindow) 
-            factory.Create(new ErrorService(), mainWindow);
+
+        if (window is MainWindow mainWindow)
+            factory.Create<IError, ErrorService, MainWindow>(new ErrorService(), mainWindow);
         else
             throw new Exception("Window not found");
     }
