--- src/GameRandom/Scr/DI/InjectAttribute.cs
@@ -1,6 +1,0 @@
-﻿namespace GameRandom.SteamSDK.DI;
-
-public class InjectAttribute
-{
-    
-}
