+++ src/GameRandom/Scr/LobbySystem.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Avalonia.Threading;
 using GameRandom.DataBaseContexts;
-using GameRandom.SteamSDK.DI;
+using GameRandom.Scr.DI;
 using Microsoft.EntityFrameworkCore;
 using Steamworks;
 
@@ -79,7 +79,7 @@ public class LobbySystem : ILobbyService
         if (callback.m_EChatRoomEnterResponse != (uint)EChatRoomEnterResponse.k_EChatRoomEnterResponseSuccess ||
             SteamManager.Instance == null)
         {
-            var error = Di.Container.GetInstance(typeof(IError)) as IError;
+            var error = Di.Container.GetInstance<IError>() as ErrorService;
             error?.ShowErrorWindow("Not found room");
             return;
         }
